requirements.txt
coqui-tts
scenedetect[opencv]
pylint
ollama
openai-whisper

torch
torchvision
pillow
git+https://github.com/openai/CLIP.git